+++ src/main.cpp
@@ -1,5 +1,5 @@
 #include "TaskManager.h"
-#include "model/SequentIdGenerator.h"
+#include "id_generation/SequentIdGenerator.h"
 #include "states/factory/LazyStatesFactory.h"
 #include "DefaultModelController.h"
 #include "user_interface/UserInterface.h"
